Avatar moves (stuck at top). BFS implemented.
While map is not complete, returned direction is next coordinate in the stack of the BFS algorithm. However, the current implementation deletes the queue with every call to calculateRoute(). This results in the avatar getting stuck back and forth when it reaches the top of the screen. This must be changed. Plan: go back to n
ot clearing the stack, and instead, checking if the next coordinate in the queue is adjacent to the current position of the avatar. If not, remove from queue, and check the next one